230914/yoda bulk 처리 스크립트 완료
