+++ sh/tmdb_pyspark.sh
@@ -12,4 +12,4 @@ MOVIE_CODE="$3"
 # Spark Job 제출
 # 각자의 spark 환경에 맞게 설정 필요
 # : 절대경로, master 링크, executor 메모리 및 코어
-$SPARK_HOME/bin/spark-submit --master spark://master:7077 --total-executor-cores=2 $SPARK_FILE $DATE $MOVIE_CODE
+$SPARK_HOME/bin/spark-submit --master spark://master:7077 --total-executor-cores=2 $SPARK_FILE $DATE
